Newest version (should work...)
Cleaned up. Removed a few tiers of polling and put interrupts in to replace them (no more timerExp, no more rx polling). The only other resource hogs that need to be dealt with are the tx polling and the interface to the phy (replace loop with a memcopy, whenever the PHY gets fixed to allow that)

git-svn-id: http://warpproject.org/svn/WARP@86 dd934ca3-7c2e-414f-9084-136bfdd87501